+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.1.100] - 2025-05-26
+
 ### Fixed
 
 - Fixed the price table not getting in to excel.